requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
blas=1.0=openblas
brotli=1.0.9=h1a28f6b_7
brotli-bin=1.0.9=h1a28f6b_7
brotli-python=1.0.9=py312h313beb8_7
bzip2=1.0.8=h80987f9_5
ca-certificates=2024.3.11=hca03da5_0
certifi=2024.2.2=py312hca03da5_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
contourpy=1.2.0=py312h48ca7d4_0
cycler=0.11.0=pyhd3eb1b0_0
expat=2.5.0=h313beb8_0
ffpyplayer=4.5.1=pypi_0
filelock=3.13.1=py312hca03da5_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h1192e45_0
giflib=5.2.1=h80987f9_3
idna=3.4=py312hca03da5_0
jinja2=3.1.3=py312hca03da5_0
jpeg=9e=h80987f9_1
kiwisolver=1.4.4=py312h313beb8_0
lcms2=2.12=hba8e193_0
lerc=3.0=hc377ac9_0
libbrotlicommon=1.0.9=h1a28f6b_7
libbrotlidec=1.0.9=h1a28f6b_7
libbrotlienc=1.0.9=h1a28f6b_7
libcxx=16.0.6=h4653b0c_0
libdeflate=1.17=h80987f9_1
libexpat=2.5.0=hb7217d7_1
libffi=3.4.4=hca03da5_0
libgfortran=5.0.0=11_3_0_hca03da5_28
libgfortran5=11.3.0=h009349e_28
libjpeg-turbo=2.0.0=h1a28f6b_0
libopenblas=0.3.21=h269037a_0
libpng=1.6.39=h80987f9_0
libsqlite=3.45.2=h091b4b1_0
libtiff=4.5.1=h313beb8_0
libwebp=1.3.2=ha3663a8_0
libwebp-base=1.3.2=h80987f9_0
libzlib=1.2.13=h53f4e23_5
llvm-openmp=14.0.6=hc6e5704_0
lz4-c=1.9.4=h313beb8_0
markupsafe=2.1.3=py312h80987f9_0
matplotlib=3.8.3=py312h1f38498_0
matplotlib-base=3.8.3=py312ha6faf65_0
mpmath=1.3.0=py312hca03da5_0
munkres=1.1.4=py_0
ncurses=6.4=h313beb8_0
networkx=3.1=py312hca03da5_0
numpy=1.26.4=py312h7f4fdc5_0
numpy-base=1.26.4=py312he047099_0
opencv-python=4.9.0.80=pypi_0
openjpeg=2.3.0=h7a6adac_2
openssl=3.2.1=h0d3ecfb_1
packaging=23.2=py312hca03da5_0
pillow=10.0.1=py312h3b245a6_0
pip=23.3.1=py312hca03da5_0
pyparsing=3.0.9=py312hca03da5_0
pysocks=1.7.1=py312hca03da5_0
python=3.12.2=hdf0ec26_0_cpython
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.12=4_cp312
pytorch=2.2.2=py3.12_0
pyyaml=6.0.1=py312h80987f9_0
readline=8.2=h1a28f6b_0
requests=2.31.0=py312hca03da5_1
setuptools=68.2.2=py312hca03da5_0
six=1.16.0=pyhd3eb1b0_1
sympy=1.12=py312hca03da5_0
tk=8.6.13=h5083fa2_1
torchvision=0.17.2=py312_cpu
tornado=6.3.3=py312h80987f9_0
typing_extensions=4.9.0=py312hca03da5_1
tzdata=2024a=h04d1e81_0
urllib3=2.1.0=py312hca03da5_1
wheel=0.41.2=py312hca03da5_0
xz=5.4.6=h80987f9_0
yaml=0.2.5=h1a28f6b_0
zlib=1.2.13=h53f4e23_5
zstd=1.5.5=hd90d995_0
